[feature] Add show.js for the upvote/downvote functionality
Implement the two methods that handle the user actions pertaining to upvote/downvote functionality.

The two methods are:

- up_vote(): invoked when a user clicks on the upvote button. The method activates the 'thumbs up'
 icon, increments the number of up votes counter and disables the 'thumbs down' icon. clicking again on
this icon decrements the upvote counter, re-enables the thumbs down icon and deactivates the thumns up icon.

- down_vote(): invoked when the user clicks on hte down vote(thumbs down) icon. The method activates
 the 'thumbs down' icon, increments the down vote counter and disables the 'thumbs up' icon. Clicking
on the icon again deactivates the icon, decrements the down vote counter and enables the 'thumbs up' icon.