[feature] Add a default_org_create action, and trigger it the first time a datarequest is created
The default organization is created only if no organization with the name provided in the production.ini file
exists.

We also supply the id of that organization to the datarequest object that is created in datarequest_create

This ensures that every datarequest is associated with the default organization, regardless of whether the
user chose an organization in the UI dropdown menu when creating the datarequest, or not.